feat(core): allow multiple configs
Before when you created more than one config you would get multiple
loaders from the previous config. Basically we are just resetting the
loaders every time the `initialize` function is called. We are now also
cloning the `webpackConfig` object so we can return a copy of the
object and not manipulate it when creating the second config.